refactor(models structure): Resolve module import issue by renaming and reorganizing file structure
- Add app/models/__init__.py to treat models directory as a package.
- Rename models.py to base.py and relocate it under app/models/ directory to avoid name conflict.
- Update import paths in app/models/dao/users.py and app/routers/users.py to reflect the new structure.
- The changes address the issue where the duplicate models name caused module import errors, improving project organization and avoiding future conflicts.
